refactor(surge-core): split artifact_contract.rs into module
Pure refactor — no behavior change, no public API change, no
serialization change. Splits the 1616-line `artifact_contract.rs`
into a module with submodules organized by artifact kind so the
file can keep growing as new kinds and stricter semantics arrive
without ballooning a single file.

Layout under `crates/surge-core/src/artifact_contract/`:

- `mod.rs` — public API facade: validate_artifact*, dispatch,
  re-exports.
- `contract.rs` — ArtifactKind, ArtifactContract, ArtifactContractRef,
  ArtifactFormat, SchemaVersionOwner, ARTIFACT_SCHEMA_VERSION,
  contract constants, all_contracts, contract_for,
  schema_version_for_kind.
- `diagnostic.rs` — ArtifactDiagnosticSeverity/Code,
  ArtifactValidationDiagnostic/Report/Error.
- `path.rs` — normalize_path, is_adr_path, is_story_path.
- `parse.rs` — shared TOML/markdown parsing and section helpers.
- `kinds/{description,requirements,roadmap,roadmap_patch,spec,adr,
  story,plan,flow}.rs` — per-kind validators and their tests.

All external imports (`surge_core::artifact_contract::*`) continue
to work unchanged via `pub use` in `mod.rs`.
