+++ crates/surge-core/src/artifact_contract/kinds/mod.rs
@@ -1,0 +1,11 @@
+//! Per-kind artifact validators.
+
+pub(super) mod adr;
+pub(super) mod description;
+pub(super) mod flow;
+pub(super) mod plan;
+pub(super) mod requirements;
+pub(super) mod roadmap;
+pub(super) mod roadmap_patch;
+pub(super) mod spec;
+pub(super) mod story;
